requirements.txt
python-dotenv
groq